requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.4
paho-mqtt==2.0.0
pandas==2.2.1
protobuf==5.26.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
tdqm==0.0.1
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
